Animation Component done!

+++ WizWorld/Include/Game.h
@@ -14,30 +14,29 @@ public:
 	void run();
 
 private:
-	/*=== Attributes ===*/
 	sf::RenderWindow m_window;
 	bool m_fullscreen;
+
 	sf::Event m_event;
+
 	std::stack<State*> m_states;
-	/* Clock*/
+
 	sf::Clock m_clock;
 	/* Delta time, time between 2 frames */
-	double m_dt;
+	float m_dt;
+
 	/* Keys that Game uses */
 	std::map<std::string, int> m_keys;
 
 
-	/*=== Private methods ===*/
+	/* Private functions */
 
 	void update();
 	void updateEvents();
 	void updateDt();
 	void render();
-	/* Initializes @member[window] with the parameters in the files "window.cfg" */
 	void initializeWindow();
-	/* Initializes @member[states] */
 	void initializeStates();
-	/* Initializes @member[keys] */
 	void initializeKeys();
 };
